Refactor: dynamics implementations
Moves dynamics implementations from `models/protocols` to `dynamics/state_space_dynamics`.
Renames `AbstractModel` to `StateSpaceDynamics` to better reflect its purpose.
This change aims to improve the code structure and organization by grouping dynamics-related classes in a dedicated directory.
