+++ bioptim/__init__.py
@@ -219,7 +219,7 @@ from .models.biorbd.model_dynamics import (
 from .models.protocols.biomodel import BioModel
 from .models.protocols.holonomic_constraints import HolonomicConstraintsFcn, HolonomicConstraintsList
 from .models.protocols.stochastic_biomodel import StochasticBioModel
-from .models.protocols.abstract_model_dynamics import (
+from .dynamics.state_space_dynamics import (
     TorqueDynamics,
     StochasticTorqueDynamics,
     HolonomicTorqueDynamics,
@@ -231,7 +231,7 @@ from .models.protocols.abstract_model_dynamics import (
     MusclesDynamics,
     JointAccelerationDynamics,
 )
-from .models.protocols.abstract_model import AbstractModel
+from .dynamics.state_space_dynamics import StateSpaceDynamics
 from .optimization.multi_start import MultiStart
 from .optimization.non_linear_program import NonLinearProgram
 from .optimization.optimal_control_program import OptimalControlProgram
