+++ bioptim/dynamics/state_space_dynamics/torque_dynamics_variational.py
@@ -1,0 +1,44 @@
+from ..configure_variables import States, Controls
+from ..dynamics_evaluation import DynamicsEvaluation
+from ..configure_variables import ConfigureVariables
+from .abstract_dynamics import StateSpaceDynamics
+
+
+class VariationalTorqueDynamics(StateSpaceDynamics):
+
+    def __init__(self):
+        super().__init__()
+        # If the model has no holonomic constraint, there will be no lambdas defined
+        # TODO: QDOT are declared as parameters in VariationalOptimalControlProgram, these parameters should instead be
+        # declared here as self.parameter_configuration = [Parameter.QDOT_INIT, Parameter.QDOT_END]
+        self.state_configuration = [States.Q, Controls.LAMBDA]
+        self.control_configuration = [Controls.TAU]
+        self.algebraic_configuration = []
+        self.functions = [
+            lambda ocp, nlp: ConfigureVariables.configure_variational_functions(ocp, nlp),
+        ]
+
+    def dynamics(
+        self,
+        time,
+        states,
+        controls,
+        parameters,
+        algebraic_states,
+        numerical_timeseries,
+        nlp,
+    ):
+        # For variational integrator, the traditional dynamics is skipped, but its declaration is mandatory in bioptim
+        return DynamicsEvaluation(
+            dxdt=nlp.cx(
+                nlp.states.shape,
+            ),
+            defects=nlp.cx(0),
+        )
+
+    def get_rigid_contact_forces(self, time, states, controls, parameters, algebraic_states, numerical_timeseries, nlp):
+        return
+
+    @property
+    def extra_dynamics(self):
+        return None
