+++ tests/shard1/test_dynamics.py
@@ -25,7 +25,7 @@ from bioptim import (
     JointAccelerationBiorbdModel,
     States,
     Controls,
-    AbstractModel,
+    StateSpaceDynamics,
 )
 
 from ..utils import TestUtils
@@ -1089,14 +1089,14 @@ def test_joints_acceleration_driven(cx, phase_dynamics):
 @pytest.mark.parametrize("contact_types", [(), [ContactType.RIGID_EXPLICIT]])
 def test_custom_dynamics(contact_types, phase_dynamics):
 
-    class CustomModel(BiorbdModel, AbstractModel):
+    class CustomModel(BiorbdModel, StateSpaceDynamics):
         def __init__(self, model_path, contact_types):
             BiorbdModel.__init__(
                 self,
                 model_path,
                 contact_types=contact_types,
             )
-            AbstractModel.__init__(self)
+            StateSpaceDynamics.__init__(self)
 
             self.state_configuration = [States.Q, States.QDOT]
             self.control_configuration = [Controls.TAU]
